feature(svelte): support svelte:component
