+++ packages/core/src/parsers/angular.ts
@@ -28,7 +28,7 @@ const transformBinding = (
   _options: AngularToMitosisOptions,
 ) => {
   return babelTransformCode(binding, {
-    Identifier(path: babel.NodePath<babel.types.Identifier>) {
+    Identifier(path) {
       const name = path.node.name;
 
       if (
